feat : 첫번째 과제 Tool interface 추가
*Tool interface 추가
*GdscMember에 의존성 주입
* pc, labtop implements Tool
