--- src/main/java/com/gdsc/homework/firstHW/HomeWork.java
@@ -1,8 +1,0 @@
-package com.gdsc.homework.firstHW;
-
-public class HomeWork {
-    public static void main(String[] args) {
-        GdscMember gdscMember = new GdscMember();
-        gdscMember.playCoding();
-    }
-}
