--- src/main/java/com/gdsc/homework/firstHW/Laptop.java
@@ -1,7 +1,0 @@
-package com.gdsc.homework.firstHW;
-
-public class Laptop {
-    public void coding(){
-        System.out.println("노트북으로 코딩하고 있어요!");
-    }
-}
