--- src/main/java/com/gdsc/homework/firstHW/PC.java
@@ -1,7 +1,0 @@
-package com.gdsc.homework.firstHW;
-
-public class PC {
-    public void coding(){
-        System.out.println("PC로 코딩하고 있어요!");
-    }
-}
