+++ src/main/java/com/gdsc/homework/firstHW/model/Tool.java
@@ -1,0 +1,5 @@
+package com.gdsc.homework.firstHW.model;
+
+public interface Tool {
+    void coding();
+}
